backend/requirements.txt
Flask
Flask-Cors
openai
pyyaml
python-dotenv
sentry-sdk>=1.30.0